musicxml2ly: Support nested tuplets
To make this work, the new algorithm in function `group_tuplets` now always
uses the elements `<tuplet-actual>` and `<tuplet-normal>` (if present) to
set a tuplet's scaling factor.  This has the side effect that exotic tuplet
number values not actually related to the real tuplet scaling cause strange
scaling values in the created `.ly` file to compensate it (the printed
output is the same as before).

Also extend the corresponding regression test.

Fixes issue #4434.