musicxml2ly: Rewrite `<direction>` handling (1/12): top-level function
This commit is the first of a series to improve the handling of
`<direction>` elements and of its `<direction-type>` children in particular.

The previous code couldn't correctly handle combinations of
`<direction-type>` children (with the exception of metronome marks, at least
to a certain extent); for example, MusicXML code for 'p subito' was resolved
into `-\p -"subito"`, which puts the two objects at the same vertical
position instead of aligning them horizontally.

Since the necessary changes are large, we start with commits that introduce
new (but inactive) functions and classes, usually ending with `_new` or
`New`, respectively.  A final step then removes all old functions, removes
the 'new' suffixes of the new functions, and activates them.

Function `musicxml_direction_to_lily_new` is the new top-level function to
handle `<direction>`.  It looks for various combinations of
`<direction-type>` children and eventually calls handlers for them
(collected in dictionary `state_dict`).  Additionally, it handles MusicXML's
'carry-over' mechanism of attributes, which exists only for
`<direction-type>`.  Handling of single, not-combined `<direction-type>`
children is very similar to the old function, `musicxml_direction_to_lily`.