musicxml2ly: Rewrite `<direction>` handling (2/12): `<metronome>`
Here comes the first function of the `<direction-type>` handlers to process
metronome marks, based on `musicxml_metronome_to_ly`.  Processing of the
optional text before the metronome mark is delegated to a new `TempoMarkNew`
class, to be introduced in a follow-up commit.  There are also two small
improvevents: The new code can handle multiple successive `<beat-unit-dot>`
elements, and it properly rejects `<beat-unit-tied>` as being 'complex'
(which isn't supported yet).

The new classes in `musicxml.py` have been added to allow tests with
`isinstance()`.