musicxml2ly: Rewrite `<direction>` handling (6/12): spanners
LilyPond needs information on the right part of a spanner before the left
part gets emitted.  This commit replaces the previous code with a more
generic solution that links the `Music_xml_spanner` node of the left part
with the right part (and vice versa), thus enabling access to both parts in
the `musicexp` module.