musicxml2ly: Add color support for many elements of `<note>`
The color inheritance for `<note>` is implemented, too, with the exception
of the following elements: `<breath-mark>`, `<caesura>`, and dynamics.  It
is questionable whether inheritance makes actually sense, given that those
objects can be far away from the note itself.

There is also no color inheritance for spanners like a glissando or a
two-stem tremolo; again, it is questionable whether this makes sense.  See
https://github.com/w3c/musicxml/issues/24 for more.

Fixes issue #3486.