musicexp.py: Activate `pre_chord_ly()`
Most of the infrastructure for calling this function was already there, but
it stayed unused for unknown reasons.

Function `pre_chord_ly` adds stuff before a chord (or a single note).  This
is in contrast to `pre_note_ly`, which adds stuff that must precede a note
directly (to make it work within a chord, too).