Rewrite handling of tied tab note heads
If a tab note head is right bound of a tie (or has a repeat tie)
and a spanner starts, always print it parenthesized.
If a Tie is split, print the tied tab note head due to the
settings of tied-properties.
Same for a tied tab note holding a `RepeatTie`.

This was already intended by the first implementation
by 3417b5edd0596f128bc5faca6c600c3b11915709
Though, using break-visibility for a note head made no sense.

dcafb42fa6c1cf234a8916ec221a2b006d3582b9 deleted break-visibility,
but introduced a regression.

This patch rewrites most of the functionality:
- extend the engraver, port it to scheme
- introduce a new context-property tabFullNotation
- introduce new sub-properties ties/repeat-tied, for
  TabNoteHead.details.tied-properties.
  Delete TabNoteHead.details.repeat-tied-properties.
- Move functionality from (Repeat)Tie.after-line-breaking to
  TabNoteHead.after-line-breaking.
- Extend tab-note-head interface
- Adjust hide/showSplitTiedTabNotes

Closes #6733