elisp: Change all prefixes to lowercase
This patch changes all the Elisp functions and symbols belonging to
the lilypond-mode.el package to lowercase. This was done to clean up
the codebase and have it follow the naming convention in Emacs, most
importantly, making the major mode name the same as the package name.

As a consequence, this will likely break the existing user config,
for example, mode hooks. They can be rewritten very quickly, however,
we need to notify the users in the release notes.

Fixes #6740.