Eliminate grob-property 'display-cautionary
Let 'parenthesize do the job.

Adds another exception in parenthesis-engraver.cc for TabNoteHead,
to avoid creating a Parentheses.
In TabStaff this will lead to the need to enter
  \parenthesize NoteColumn
if it is wished to parenthesize the whole chord, otherwise the
TabNoteHeads would be parenthesized one for one.
Imho, semantically more correct and extends the possibilities of
\parenthesize.
Regtest parenthesize-chords.ly is amended.
NR is updates.

Closes #4267