Partcombine: simple alist lookup for make-part-combine-marks default
State machine traversal is equivalent to alist lookup as long as the value for
a given key is the same for all machine states (which is indeed the case for the
default part-combine-marks state machine table).

Added a public alist equivalent to the default state machine.
Added lookup function suitable to pass into make-part-combine-marks.
These are now the default used by make-part-combine-marks, since alist lookup
is more efficient.

The state machine table and traversal function are left as a template for future
customization.