Hara_kiri_group_spanner: ignore system bound columns for pure suicide
During pure positioning, assume that grobs belonging only to the leftmost or
rightmost column are never worth-living. The left and right column ranks passed
to request_suicide_alone *should* always be non-musical columns considered as
potential system bounds. Note that since consider_suicide now calls the unpure
methods, this change to request_suicide_alone only affects pure positioning and
line breaking, not whether staves are actually hidden.

The effect of this change is hard to judge concretely, but in principle, except
for the case where break-visible non-musical items are genuinely worth living
(which I suspect is rare), it should keep pure positioning assumptions in line
with the unpure behavior.