Improve octavation spanner MusicXML tests
Due to unfortunate historical circumstances there are two incompatible ways
to position end tags of spanners that are encoded via the `<direction>`
elements in the stream of MusicXML elements, namely before the anchor note
(as done in files exported by Finale or Sibelius) or after (as done by
MuseScore, for example).  Especially crucial are ottava lines, because they
are also responsible for altering the displayed octavation of the affected
notes as a side effect.

* Use smaller, more sensible values for `<offset>`.
* Rename the 'after' file and add a 'before' file.