measure_position always uses euclidean_remainder
Add overloads of measure_position () so that it may be called with
supplied position and length arguments instead of looking them up in the current context.  This can be useful to reduce property lookups when the
caller already knows these values, or to calculate with past values or
hypothetical values.

measure_position () now always finds the remainder modulo the measure
length whether the raw position is positive or negative.  Before, it
was not likely that the position would exceed the length; however, it
could sometimes happen given how \partial was implemented until
recently, and of course the situation could still be forced with \set
commands.  This is therefore a little more robust for things that expect
a value in [0, measureLength).