musicxml2ly: Work around LilyPond's issue #34 (2/6)
Collect grace note sequence lengths of all voices at the beginning of a
piece in a dictionary and pass the difference to the longest one to
`musicxml_voice_to_lily_voice`.

This commit doesn't handle grace note synchronization at other musical
moments (which is left as a TODO for the moment).