Fix easyHeads with magnifyStaff
Previously note-head::brew-ez-stencil looked at staff-space and
font-size to scale the note head accordingly. Though, magnifyStaff
changes both and note-head::brew-ez-stencil fails in this case.
The fix is simply to not scale with the alread scaled staff-space,
but to use the staff-space from grob's layout.

Regressiontest easy-notation-size.ly is amended.
Due to issue #6756 the output is partly not correct.

Closes #6755