check_texi_refs.py: Fix file searching algorithm
After re-ordering the `Documentation` subdirectory tree a few years ago
there is only a single `out-www` directory in it if we have an in-tree build
(as needed by `check_texi_refs.py`).  Since the subdirectory tree of this
`out-www` directory has the same layout as the `Documentation` subdirectory
tree itself we have to use `outdir` as the first element in constructing a
search path, not as the second.

Also normalize relative paths to improve legibility of warnings.