baseMoment->beatBase in timeSignatureSettings
The type of this alist entry is a number, not a moment, so the name
baseMoment was confusing.

No convert-ly rule is implemented because baseMoment is also the name of
a context property which is actually a moment.  Hopefully, foregoing
conversion will impact a smaller number of users than a tricky (probably
sloppy) conversion would, and the users this does impact will appreciate
the change.

Update the documentation of \overrideTimeSignatureSettings arguments in
the English NR.  Add `#` to the beatBase value for consistency with its
internal type, though the function still accepts a fraction.

Update just examples and argument names in languages other than English.