musicxml2ly: Add `single_voice` property to `Note` and `Rest` classes
We need this for better ad-hoc polyphony support that often switches between
single-voice and multi-voice parts (to be implemented in the next commit).

The new code uses a double-linked list – we need deletion of elements while
walking over the list repeatedly.  Since Python doesn't support this
natively, the relevant code from the 'llist' Python package (see
https://github.com/rgsoda/pypy-llist for the repository) is imported, too.