Fix a corner case in calc-measure-length
(ly:make-moment a b) creates a moment with main part a/b when b > 0, but
it creates a moment with main part a/1 when b = 0.  The discontinuity is
not desired in calc-measure-length.

This commit causes a time-signature fraction with a zero denominator to
set measureLength to +inf.0.  The change is evident in regression test
time-denominator-zero, where a bar line has disappeared because the
measure does not end.