lilypond-book: Improve DocBook input
Contrary to other formats, `lilypond-book` doesn't introduce a new command
in 'DocBook' mode for LilyPond input.  Instead, it re-uses the
`<programlisting> ... </programlisting>` block command (with a
`language="lilypond"` attribute).  However, this command expects that its
data is properly escaped, that is, characters `>`, `<`, and `&` must be
encoded as `&gt;`, `&lt;`, and `&amp;`, respectively.

This commit properly 'unescapes' the data before feeding it to LilyPond.

Also add a regression test.

Fixes issue #6204.