Improve \popContextProperty underflow warning
The warning is now tied to the source location.  The text of the warning
is simpler because the property path is clear from the source.

The warning had said that the property would be set to its default
value, but the code actually unsets it.  The warning has been reworded.

Test coverage of the warning is moved into a new case that shows a
visible consequence of unsetting the property.