doc: Rename 'LilyPond index' to 'Index'
We now have only a single index.

Note that the index node itself is called differently: this is for the sake
of the split HTML documentation, which contains `@node index` for the
top-level page of a manual so that the corresponding HTML page is called
`index.html`, as is common usage.