Remove the value arg from \pushContextProperty
\pushContextProperty now does nothing more than save the current value.

This makes it possible to save the current value without changing it
immediately (or knowing the current value and setting it unnecessarily).

This also eliminates a wasteful set in the case of pushing and unsetting
a property.  Before, that would have required this:

    \pushContextProperty foo unnecessaryValue
    \unset foo

Now, it takes just this, which parallels the push-then-set case:

    \pushContextProperty foo
    \unset foo