doc: fix generation of snippet pages
Commit 3eb5d20e3 introduced node prefixes to allow referencing of individual
snippets.  (Node prefixes are necessary because some snippets appear more
than once in the 'snippets' document.)

However, this change used language-specific translations of the snippet
titles for the construction of node names, which caused two unwanted side
effects.

* `texi2any --html` converts non-ASCII characters to ASCII-only
  transliterations in links (see
  https://lists.gnu.org/archive/html/bug-texinfo/2025-02/msg00025.html for a
  discussion of problems related to its default behaviour).  As a result,
  some HTML pages have overlong and ugly names and element IDs like

    https://lilypond.org/doc/v2.25/Documentation/snippets/\
      pitches_003a-yin-yu-woboisugotonizhui-jia-suru\
      #pitches_003a-_97f3_57df_3092_30dc_30a4_30b9_3054_3068_306b_8ffd_52a0_3059_308b

  for a Japanese translation.

* More seriously, it completely disabled the language switching mechanism in
  HTML pages (i.e., the 'Other languages:' line in the footer was missing).

This commit fixes both problems.  It modifies `lys-to-tely.py` to emit
`@node` lines, always based on the English snippet titles.  To get these
titles, option `-I` is added to the make rule so that the snippet files are
actually found.

The Texinfo macro setup is adapted to accommodate for the changes.