Number-over-note time signatures
Setting the denominator-style property to 'note prints denominators as a
note and dots when exact representation is possible.

Properties note-staff-position and note-dots-direction control the
placement of the note.

Properties note-head-style and note-flag-style control the appearance.

All of these are available in both the TimeSignature grob interface and
the \compound-meter markup command.