*/changes.tely: Restore top-level `@menu` block
It is a feature of `texi2any` to also generate a `@detailmenu` block if the
top-level menu gets auto-generated, which we don't want.

As soon as we are going to use Texinfo version 7.3 or newer this workaround
can be replaced by setting the `FORMAT_MENU` option in file `lilypond.init`
to value `menu_no_detailmenu`.