Fix double announcement of SpanBarStub grobs
Span_bar_stub_engraver was calling announce_grob (Grob_info, Context *).
It was reasonable to think that this would announce the grob in the
given context and enclosing contexts, but what the function actually did
was announce the grob in the given context and enclosing contexts AND
again in the local context.  Any engraver listening for span bars in the
same context as Span_bar_stub_engraver would acknowledge each one twice.

This commit changes announce_grob (Grob_info, Context *) so that it no
longer announces the grob a second time in the local context.

The other caller of announce_grob (Grob_info, Context *) was
Auto_beam_engraver, which actually relied on the old behavior.  It now
directly calls announce_grob_locally_only () for clarity.