musicxml2ly: Fix handling of `<group-barline>`
In LilyPond, bar lines within a `StaffGroup` grob or its siblings are
connected by default.  However, the MusicXML default is to *not* connect bar
lines; it is necessary to have explicit `<group-barline>` objects to switch
on bar line connections.  Additionally, if a staff group is set up to
connect bar lines, it is not possible to change this decision in a nested
group.

`musicxml2ly` used `\hide SpanBar` to switch off *all* bar line connections
within a `StaffGroup`.  This didn't work correctly because there was no code
to un-hide the `SpanBar` grob within nested groups.

This commit replaces the use of `\hide SpanBar` with the more fine-grained
'\override Staff.BarLine.allow-span-bar = ##f` to fix the issue.