lilypond-book: Emit `.ly` link for `.xml` files
Between LilyPond version 2.14 and 2.15, the link of both the file name (if
option `printfilename` is specified) and the image changed from `.ly` to
`.xml` in `musicxml/collated_files.html`.  This commit changes the image
link back to `.ly` so that both the source MusicXML files and the
intermediate LilyPond files can be inspected.