create-weblinks-itexi.py: Better link description for PDFs
[This is for the website build only, not for the offline build.]

Assume you want to download the Catalan version of the Usage Manual (to name
an example).  If you click on the 'details of Usage' link and change to the
Catalan version, you see the following:

```
Ús (HTML seccionat) ...
Ús (HTML monolític) ...
Ús.pdf ...
```

However, the link `Ús.pdf` points to a file called `usage.ca.pdf`, which is
misleading.  With this commit you get

```
Ús (HTML seccionat) ...
Ús (HTML monolític) ...
Ús (PDF) ...
```

instead.