create-weblinks-itexi.py: Provide locale-specific stub macros
Some LilyPond manuals will stay untranslated (like the CG), some manuals are
not available in all formats (for example, there are no PDF manuals for
Japanese).  However, there was no support for such a situation with the
available macros in `weblinks.itexi`; it was only possible to use the
English versions regardless of the language, causing English descriptions
within non-English locales.

This commit introduces 'stub' macros that use locale-specific translations
but point to the English version of the document.