Add support for auto-generated lyric extenders.
If 'autoExtenders' is set to #t, LilyPond will generate lyric
extenders for any lyric syllable that occurs in a melisma
and is not followed by ' -- '.  By default, these extenders
will be suppressed if they would be too short, but this
behaviour can be switched off by overriding
'LyricExtender.remove-short-autoextender'.

The behaviour for manually lyric extenders entered by the user
via ' __ ' is not changed.

Internally, the approach is to create a tentative LyricExtender
grob for each lyric syllable and killing it if it turns
out not to be needed.

Closes #4509.