Fix shortening of ledger lines with an accidental
`Ledger_line_spanner::print` now always shortens the correct ledger line
if a note head has an accidental; it no longer implicitly assumes that the
first element of the ledger position list is the one to change (this might
not be the case if `ledger-extra` is set).

While we are at it, add code to shorten more ledger lines, depending on
the accidental.

Also update regression test.