Make `ledger-positions` sublists really work
The condition ending the loop looked at the last processed sublist element.
However, this element might not be in the ledger line interval, causing the
loop to stop too early.

We now store the positions in a temporary array and merge only with the
`values` array if at least one of them is actually in the ledger line
interval.