Canonical time signature representation
Establish a canonical representation for time signatures, where a
fraction is consistently represented by a pair with a number in the cdr,
and concatenation is consistently represented with a list.

Example: simple `'(10 . 8)` and subdivided `'((3 3 2 2) . 8)`.

This will simplify many internal consumers of time signatures.
A function that reads a canonical time signature does not have to handle
multiple representations of the same thing (e.g., `'(4 . 4)` and
`'(4 4)`) with context-dependent interpretations of lists and different
ways of getting the denominator.

The `\compoundMeter` command still accepts the old abbreviated
representation (i.e., `'(3 3 2 2 8)` for the subdivided example above).
The process of expanding the abbreviation improves the validation of
this argument.