NR: Use `@part` (1/2)
The NR needs a lot of sectioning levels, going down to level five (i.e.,
`@subsubsubheading`), which LilyPond must specifically define because
Texinfo only supports four levels in general.  To avoid long four-part
sectioning numbers like '1.4.5.2', level-four and level-five sections are
unnumbered in the NR; however, some of the level-four sections are quite
long, and it would be thus beneficial if they had also sectioning numbers so
that they could be referenced more easily.

This commit is a first step to remedy the situation.  It replaces `@chapter`
with `@part` and lifts all other sectioning commands by one level.