musicxml2ly: Better support for `<credit>`
This commit changes how `musicxml2ly` handles `<credit>` elements: if at
least one is present, metadata elements (`<work>`, `<identification>`, etc.)
are no longer used to fill up LilyPond's standard fields in the `\header`
block.  The are two reasons to do that.

* `<credit>` data has formatting attributes like the font size or the
  enclosure type, which comes nearer to what a user sees on screen or paper
  before exporting to MusicXML.  `musicxml2ly` can now handle these
  attributes.

* Contrary to other applications that rely on exact positioning of objects,
  `musicxml2ly` only uses the coordinates of `<credit>` children for
  comparison purposes (namely to guess a proper `\header` field if no type
  is present); it doesn't have to adjust them in any way.

[Using `--credit-page=0` it is possible to make `musicxml2ly` ignore
 `<credit>` elements.]

The logic to handle `<credit>` elements has been improved, too: We no longer
rely on the order of its children while mapping to `\header` fields.
Instead, we do a first pass to get the types so that we can prefer the
elements with real types over the ones with guessed types.

All elements that are not mapped to standard `\header` fields (and which are
not ignored like the 'page number' type) are still output.  To do that in a
consistent manner, `musicxml2ly` uses the prefix 'id: ' for
`<identification>` children, 'credit: ' for `<credit>` elements, and the
MusicXML element name (without a prefix) in all other cases.  As a
consequence, some (unused) fields got new names.

    movementnumber       →  movement-number
    encodingsoftware     →  id: software
    encodingdate         →  id: encoding-date
    encoder              →  id: encoder
    encodingdescription  →  id: encoding-description
    source               →  id: source

Also update and add regression tests.