timeSignatureFraction -> timeSignature
The new property accepts a superset of the values that the old property
accepted, so convert-ly can safely replace the name in assignments.

Other cases are left untouched because Scheme code reading
timeSignatureFraction would likely choke on subdivided or alternating
time signatures.  To support this, timeSignatureFraction remains as a
deprecated property.