musicxml2ly: Fix emitted code for elements right before a bar line
Make `add_pending_elements` differentiate between needing `<>`
unconditionally (for full-measure rests) and the case where we have elements
right before a bar line.

Also remove the insertion of `s1024` at the end of music, which was a bad
idea.  As a consequence, some elements are not displayed at the end of
music; however, a different solution has to be found anyway eventually.