binaries: Disable glib_debug
Upstream turned this on by default in version 2.82.0, but recommends
distributions to disable it. Before it was an auto-feature subject
to the generic --auto-features=disabled in our MesonPackage.

This also fixes #6836 which is caused by libgobject-2.0-0.dll linking
against libgcc_s_seh-1.dll because of the debug code.

On the other hand, libmount is now an auto-feature and therefore
disabled by default.