musicxml2ly: Fix output of `\tweak`
Use `-\tweak` if we have a grob that takes a direction.  We need this to
make `\after` work without additional braces for its second argument – it
would be a big complication if we had to insert them.  For example, this
code (lacking the `-`)

```
\after 2 \f f'1 ^\fermata \tweak color "#FF0000" ^"foo"
```

produces an `Unattached TextScriptEvent` warning, and 'foo' is omitted.