musicxml2ly: Support `<haydn>`, `<schleifer>`, `<soft-accent>`, `<spiccato>`
`<spiccato>` is a bit problematic: the glyph shown in the MusicXML standard
is a variation of a 'staccatissimo' glyph (a concave vs. a convex top as the
only difference); at smaller sizes it is not possible at all to see that
they are not the same.  Additionally, the normal way in the literature is to
use staccato dots together with the word "spicc.", primarily for teaching
materials.  Since LilyPond doesn't have different glyphs for 'spiccato' and
'staccatissimo', and MusicXML is primarily a standard to represent a score
visually and not structurally, we simply make them identical.