Remove snippet `[obsolete]-adding-a-qr-code.ly`
In the Wiki, this file has been added to a page called 'Adding a QR code'
because `[]` are not valid characters in MediaWiki page names.  However,
some changes were added so it that currently doesn't follow the expected
template used by all other snippets.  We omit it for the time being.